requirements.txt
# Ultravox MCP Server Requirements

# Core MCP and API
fastmcp>=0.6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Web Framework (optional, for HTTP wrapper)
flask>=2.3.0
flask-cors>=4.0.0

# JSON RPC
jsonrpc>=1.13.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.0.0

# Async support
aiohttp>=3.9.0
asyncio-contextmanager>=1.0.0

# Logging
python-json-logger>=2.0.7

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Testing
mock>=5.1.0
faker>=19.0.0
